Split promotion integration coverage
