+++ tests2/tests-map.json
@@ -112,7 +112,16 @@
     },
     {
       "path": "tests2/integration/session-goal-promotion.test.ts",
-      "reason": "Integration coverage for owner-scoped worktree-mode routes, authority rejection, in-place acceptance, idempotency, coordinate continuity, and proposal cleanup.",
+      "reason": "Integration coverage for promotion compensation, retries, lifecycle exclusion, exact checkout continuity, multi-repo cleanup, and sandbox ownership continuity.",
+      "execution": {
+        "runner": "vitest",
+        "tier": "unit",
+        "project": "integration"
+      }
+    },
+    {
+      "path": "tests2/integration/session-goal-promotion-validation.test.ts",
+      "reason": "Integration coverage for owner authentication, current-state workflow revalidation, optional-step precedence, and generated-default transactionality.",
       "execution": {
         "runner": "vitest",
         "tier": "unit",
